+++ package/CHANGELOG.md
@@ -1,4 +1,4 @@
-## Next
+## 0.5.0 (2026-08-22)
 
 ### Breaking Changes